Update gauss_method.F90
Gauss method worked fine but it has to be changed a little bit because of our matrix in main program, which is N - 1 x N size. We have here N, and in main program we are using gauss(..., N-1) so we have to add extra 1 to i in columns + start with j equals 0 (adding extra one as well everywhere) to have "N" (not N-1) columns. Moreover, we do not need main program created here as an example anymore, so I will make it as a comment. We also do not need extra computing for A matrix (also marked as an comment).