Update main.F90
"err" var was defined before because we have to compute the error. Instead of better name "y" I am using i variable to save some memory :). The correct function was given in the task description, so it is quite simple to compute. It needs type conversion to be correct as i and N was declared integer.